+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/model/TrackTableModel.java
@@ -39,7 +39,7 @@ public final class TrackTableModel extends SwingEntityTableModel {
   public void raisePriceOfSelected(BigDecimal increase) throws DatabaseException {
     if (selectionModel().selectionNotEmpty().get()) {
       Collection<Long> trackIds = Entity.values(Track.ID, selectionModel().getSelectedItems());
-      Collection<Entity> result = connectionProvider().connection()
+      Collection<Entity> result = connection()
               .execute(Track.RAISE_PRICE, new RaisePriceParameters(trackIds, increase));
       replace(result);
     }