+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerEditPanel.java
@@ -102,7 +102,7 @@ public final class CustomerEditPanel extends EntityEditPanel {
     @Override
     public Collection<String> get() {
       try {
-        return editModel().connectionProvider().connection().select(Customer.STATE);
+        return editModel().connection().select(Customer.STATE);
       }
       catch (DatabaseException e) {
         throw new RuntimeException(e);