+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityTablePanel.java
@@ -1279,7 +1279,7 @@ public class EntityTablePanel extends JPanel {
   private void showEntityMenu() {
     Point location = popupLocation(table);
     tableModel.selectionModel().selectedItem().ifPresent(selected ->
-            new EntityPopupMenu(selected.copy(), tableModel.connectionProvider().connection()).show(table, location.x, location.y));
+            new EntityPopupMenu(selected.copy(), tableModel.connection()).show(table, location.x, location.y));
   }
 
   private void onConditionChanged() {